role auth

+++ documents/views.py
@@ -80,12 +80,15 @@ def document_create(request):
 
 @csrf_exempt
 def documents_deleteAll(request):
-    if request.method == 'DELETE':
-        print('deleting all documents')
-        delete_all_documents()
-        return HttpResponse("All documents deleted", 'application/json')
-    else:
-        return HttpResponse("Method not allowed", 'application/json')
+    role = getRole(request)
+    if role == "Administrador":
+        if request.method == 'DELETE':
+            print('deleting all documents')
+            delete_all_documents()
+            return HttpResponse("All documents deleted", 'application/json')
+        else:
+            return HttpResponse("Method not allowed", 'application/json')
+    
     
 @csrf_exempt
 def document_delete(request, pk):
